backend/requirements.txt
Flask==2.1.0
Flask_Cors==3.0.9
openai==1.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==2.1.2
scikit-learn==1.3.0
gradio==4.27.0